requirements-prod.txt
# Core dependencies only needed for production
Flask>=2.0
flask-socketio>=5.3.6
python-socketio>=5.9.0
eventlet>=0.33.3
gunicorn>=20.0.4
PyYAML
SQLAlchemy
requests
aiohttp>=3.8 